An alternate version of board.generateNextBoard
Here is an example of how the Game object can collaborate with the
Board object more effectively.

In the previous version, the Game object knows about the internal data
structure that makes up a Board.

This version is meant to illustrate the following points:

Prefer to "tell don't ask" when collaborating with other objects.
Rather than pulling out the List from within Board and creating a
2-dimensional array the same way Board does, note how this version only
depends on two pieces of information from the Board, it's dimensions.
Rows & columns (or height & width, or other synonyms for dimension).

This enables Board to encapsulate the data structure it is using to
manage the state of the board. If, for some reason down the road, we
want to use a Set or a HashMap or a LinkedList to manage the Board, we
can now make that change without breaking the Game object. This reduces
duplication between the classes and also reduces the coupling. Game no
longer has to know about internal details of how Board is implemented.